(#3)_refactor :: view 파일 생성 및 이동
